+++ src/bin/run-once.ts
@@ -4,7 +4,7 @@ import Engine from "../lib/engine/engine";
 import { IO } from "../lib/io/io";
 import log from '../lib/log/logger';
 import { Database } from "../lib/model/database";
-import { cli } from "../lib/parameters/cli";
+import { cli } from "../lib/parameters/cli-args";
 
 main();
 async function main() {
